+++ flutter/test/android_platform_exception_event_processor_test.dart
@@ -24,15 +24,8 @@ void main() {
 
   group(AndroidPlatformExceptionEventProcessor, () {
     test('exception is correctly parsed', () async {
-      final exception = fixture.options.exceptionFactory
-          .getSentryException(testPlatformException);
-
-      final event = SentryEvent(
-        exceptions: [exception],
-        throwable: testPlatformException,
-      );
-
-      final platformExceptionEvent = await fixture.processor.apply(event);
+      final platformExceptionEvent =
+          await fixture.processor.apply(fixture.eventWithPlatformStackTrace);
 
       final exceptions = platformExceptionEvent!.exceptions!;
       expect(exceptions.length, 2);
@@ -45,6 +38,35 @@ void main() {
         "Unsupported value: '[Ljava.lang.StackTraceElement;@ba6feed' of type 'class [Ljava.lang.StackTraceElement;'",
       );
       expect(platformException.stackTrace!.frames.length, 18);
+    });
+
+    test(
+        'Dart thread is not current and not active if Android exception is present',
+        () async {
+      final platformExceptionEvent =
+          await fixture.processor.apply(fixture.eventWithPlatformStackTrace);
+
+      final exceptions = platformExceptionEvent!.exceptions!;
+      expect(exceptions.length, 2);
+
+      expect(platformExceptionEvent.threads?.first.current, false);
+      expect(platformExceptionEvent.threads?.first.crashed, false);
+    });
+
+    test('platformexception has Android thread attached', () async {
+      final platformExceptionEvent =
+          await fixture.processor.apply(fixture.eventWithPlatformStackTrace);
+
+      final exceptions = platformExceptionEvent!.exceptions!;
+      expect(exceptions.length, 2);
+
+      final platformException = exceptions[1];
+      final platformThread = platformExceptionEvent.threads?[1];
+
+      expect(platformException.threadId, platformThread?.id);
+      expect(platformThread?.current, true);
+      expect(platformThread?.crashed, true);
+      expect(platformThread?.name, 'Android');
     });
 
     test('does nothing if no PlatformException is there', () async {
@@ -62,17 +84,10 @@ void main() {
     });
 
     test('does nothing if PlatformException has no stackTrace', () async {
-      final exception = fixture.options.exceptionFactory
-          .getSentryException(emptyPlatformException);
+      final platformExceptionEvent =
+          await fixture.processor.apply(fixture.eventWithoutPlatformStackTrace);
 
-      final event = SentryEvent(
-        exceptions: [exception],
-        throwable: emptyPlatformException,
-      );
-
-      final platformExceptionEvent = await fixture.processor.apply(event);
-
-      expect(event, platformExceptionEvent);
+      expect(fixture.eventWithoutPlatformStackTrace, platformExceptionEvent);
     });
   });
 }
@@ -80,6 +95,33 @@ void main() {
 class Fixture {
   late AndroidPlatformExceptionEventProcessor processor =
       AndroidPlatformExceptionEventProcessor(options);
+
+  late SentryException withPlatformStackTrace = options.exceptionFactory
+      .getSentryException(testPlatformException)
+      .copyWith(threadId: 1);
+
+  late SentryException withoutPlatformStackTrace = options.exceptionFactory
+      .getSentryException(emptyPlatformException)
+      .copyWith(threadId: 1);
+
+  late SentryEvent eventWithPlatformStackTrace = SentryEvent(
+    exceptions: [withPlatformStackTrace],
+    throwable: testPlatformException,
+    threads: [dartThread],
+  );
+
+  late SentryEvent eventWithoutPlatformStackTrace = SentryEvent(
+    exceptions: [withoutPlatformStackTrace],
+    throwable: emptyPlatformException,
+    threads: [dartThread],
+  );
+
+  late SentryThread dartThread = SentryThread(
+    crashed: true,
+    current: true,
+    id: 1,
+    name: 'main',
+  );
 
   SentryFlutterOptions options = SentryFlutterOptions(dsn: fakeDsn);
 }